Dropped all authenticated access support
Subversion authentication has been broken for a long time and the workarounds
weren't decent. It will be back in 1.7.x.

Also added configuration option svn-command in the global section, allowing to
replace the default svn command.

And use svn+ssh:// URLs to be in BatchMode, in order to not have any
interactivity at all with ssh