Refactored log api: move specifile_svn2rpm to get_changelog
The old specfile_svn2rpm will be used only to the purpose of creating the
spec with the changelog to be used by getsrpm. The new main function to
retrieve the changelog of a given package is the function get_changelog(),
which creates the changelog from svn, append oldlog, append the changelog
from the spec (if available), sort them and create the final changelog to
be output.